Remove standard hash function constructors.
The only use-case of these was to create hash tables of the same type
but with different hash functions.  (Since for equal, string, and symbol
based hash tables, a double-hashing implementation can just derive the
pair of hash functions to use from the given canonical hash function.)
Since that's a pretty obscure use-case, and these constructors might be
annoying to define when the standard hash functions are implemented in
C, we drop these constructors.